pyproject.toml
[project]
name = "mapping_violence"
version = "0.1.0"
description = "Mapping Violence: Django application for the Mapping Violence project."
authors = ["Jason Heppler <jheppler@gmu.edu>"]
license = "MIT"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "daphne>=4.2.1",
    "django>=5.2.7",
    "django-allauth>=65.12.1",
    "django-autocomplete-light>=3.12.1",
    "django-environ>=0.12.0",
    "django-extensions>=4.1",
    "django-filter>=24.3",
    "django-import-export>=4.3.12",
    "django-schema-viewer>=0.5.3",
    "django-tables2>=2.7.0",
    "django-tailwind>=4.3.1",
    "django-unfold>=0.69.0",
    "edtf>=5.0.0",
    "geopy>=2.4.1",
    "pillow>=11.3.0",
    "psycopg2-binary>=2.9.11",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "wagtail>=7.1.2",
    "whitenoise>=6.11.0",
]

[dependency-groups]
dev = [
    "black>=25.9.0",
    "django-debug-toolbar>=6.1.0",
    "djhtml>=3.0.10",
    "pre-commit>=4.3.0",
    "pylint-django>=2.6.1",
    "pytest-django>=4.11.1",
]
